feat : detail head component 작성
- next theams를 활용해 css 처리를 위해 Header 컴포넌트 생성
