+++ apps/swifty-bank/app/agreement/_component/AgreementBtn.tsx
@@ -1,4 +1,5 @@
 "use client";
+
 import { Button } from "@swifty/ui";
 import styles from "./agreementBtn.css";
