+++ apps/swifty-bank/app/detail/page.tsx
@@ -1,7 +1,6 @@
 import { Heading } from "@swifty/ui";
-import Image from "next/image";
-import Link from "next/link";
 import styles from "./page.css";
+import Header from "./_component/Header";
 
 type TermsOfDetail = {
   title: string;
@@ -32,26 +31,7 @@ export default function DetailPage() {
   ];
   return (
     <div className={styles.container}>
-      <header
-        className={styles.header({
-          borderBottom: "light",
-        })}
-      >
-        <Link href="/agreement" className={styles.link}>
-          <Image
-            src={"/icon/Icon_Cancellation.svg"}
-            alt="Icon_Cancellation"
-            width={30}
-            height={30}
-          />
-        </Link>
-        <Heading type="h2">[필수] 개인정보 처리방침 동의</Heading>
-        <p className={styles.content}>
-          개인정보의 수집·이용에 관한 동의를 거부하실 수 있습니다. 다만, 본
-          동의는 '인증서 등록'을 위한 필수적인 사항이므로 동의를 거부하시는 경우
-          업무 수행이 불가능 할 수 있습니다.
-        </p>
-      </header>
+      <Header />
       <main>
         <ul className={styles.list}>
           {termsOfDeatils.map((detailTerm) => (
